Added Ui::Window::WithHandle_t. Main window handle closing/re-establishing working.
Side benefits:
* Got rid of some dumb virtual functions like GetSSWindowRect().
  * Probably room for more excision in this area.
* Move the "Input" and "Scroll" related methods from UI::WIndow::Base_t
  to Ui::Window::WithHandle_t.
  * Might be room for more excision here as well.

This was all in prep for making individual windows own their children.
E.g:
MainWindow owns:
  BrokerFrame
BrokerFrame owns:
  BrokerBuy
  BrokerSell
BrokerSell "owns":
  SetPricePopup

This is a sane way to do things, and should (hopefully) simplify
TrWindowType and really reduce MainWindow to an even thinner skeleton
than it already is.

Still might need to address the "MainWindow everywhere" problem, but
in fact most of the usages seem justified at this point.